+++ cesium/internal/unary/gc_test.go
@@ -104,11 +104,17 @@ var _ = Describe("Garbage Collection", func() {
 
 					Expect(frame.SeriesAt(2).TimeRange.End).To(Equal(33 * telem.SecondTS))
 					series2Data := telem.UnmarshalSlice[int64](frame.SeriesAt(2).Data, telem.Int64T)
-					Expect(series2Data).To(ConsistOf(30, 31, 32))
+					Expect(series2Data).To(ConsistOf(int64(30), int64(31), int64(32)))
 
 					Expect(frame.SeriesAt(3).TimeRange.Start).To(Equal(75 * telem.SecondTS))
 					series3Data := telem.UnmarshalSlice[int64](frame.SeriesAt(3).Data, telem.Int64T)
-					Expect(series3Data).To(ConsistOf(75, 76, 77, 78, 79))
+					Expect(series3Data).To(ConsistOf(
+						int64(75),
+						int64(76),
+						int64(77),
+						int64(78),
+						int64(79),
+					))
 				})
 			})
